+++ bplsqlparse/filters/tokens.py
@@ -2,11 +2,11 @@
 #
 # Copyright (C) 2016 Andi Albrecht, albrecht.andi@gmail.com
 #
-# This module is part of python-sqlparse and is released under
+# This module is part of python-bplsqlparse and is released under
 # the BSD License: https://opensource.org/licenses/BSD-3-Clause
 
-from sqlparse import tokens as T
-from sqlparse.compat import text_type
+from bplsqlparse import tokens as T
+from bplsqlparse.compat import text_type
 
 
 class _CaseFilter(object):