+++ bplsqlparse/keywords.py
@@ -2,12 +2,12 @@
 #
 # Copyright (C) 2016 Andi Albrecht, albrecht.andi@gmail.com
 #
-# This module is part of python-sqlparse and is released under
+# This module is part of python-bplsqlparse and is released under
 # the BSD License: https://opensource.org/licenses/BSD-3-Clause
 
 import re
 
-from sqlparse import tokens
+from bplsqlparse import tokens
 
 
 def is_keyword(value):
@@ -59,7 +59,7 @@ SQL_REGEX = {
         (r'(?<!\w)[$:?]\w+', tokens.Name.Placeholder),
 
         # FIXME(andi): VALUES shouldn't be listed here
-        # see https://github.com/andialbrecht/sqlparse/pull/64
+        # see https://github.com/andialbrecht/bplsqlparse/pull/64
         # IN is special, it may be followed by a parenthesis, but
         # is never a function, see issue183
         (r'(CASE|IN|VALUES|USING)\b', tokens.Keyword),