+++ docs/source/changes.rst
@@ -1,14 +1,14 @@
 .. _changes:
 
 ============================
- Changes in python-sqlparse
+ Changes in python-bplsqlparse
 ============================
 
 Upcoming Deprecations
 =====================
 
-* ``sqlparse.SQLParseError`` is deprecated (version 0.1.5), use
-  ``sqlparse.exceptions.SQLParseError`` instead.
+* ``bplsqlparse.PLSQLParseError`` is deprecated (version 0.1.5), use
+  ``bplsqlparse.exceptions.PLSQLParseError`` instead.
 
 
 Changelog